pyproject.toml
[project]
name = "sauron-eye"
version = "0.1.0"
description = "Сервис для получение актуальных геоданных"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "apache-airflow>=2.10.3",
    "pytest>=8.3.3",
]
